+++ app/src/main/kotlin/net/primal/android/profile/domain/ProfileDataDO.kt
@@ -15,7 +15,7 @@ import net.primal.domain.nostr.serialization.toNostrJsonObject
 import net.primal.domain.nostr.utils.decodeLNUrlOrNull
 import net.primal.domain.nostr.utils.parseAsLNUrlOrNull
 import net.primal.domain.nostr.utils.parseHashtags
-import net.primal.domain.nostr.utils.parseNostrUris
+import net.primal.domain.nostr.utils.parseNostrUriCandidates
 import net.primal.domain.profile.ProfileData
 
 fun List<NostrEvent>.mapAsProfileDataDO(
@@ -57,7 +57,7 @@ fun NostrEvent.asProfileDataDO(
         lightningAddress = metadata?.lud16,
         lnUrlDecoded = metadata?.lud16?.parseAsLNUrlOrNull() ?: metadata?.lud06?.decodeLNUrlOrNull(),
         about = metadata?.about,
-        aboutUris = metadata?.about?.let { it.detectUrls() + it.parseNostrUris() } ?: emptyList(),
+        aboutUris = metadata?.about?.let { it.detectUrls() + it.parseNostrUriCandidates() } ?: emptyList(),
         aboutHashtags = metadata?.about?.parseHashtags() ?: emptyList(),
         displayName = metadata?.displayName,
         avatarCdnImage = metadata?.picture?.let {