+++ data/caching/repository/src/commonMain/kotlin/net/primal/data/repository/mappers/remote/MetadataEvents.kt
@@ -16,7 +16,7 @@ import net.primal.domain.nostr.serialization.toNostrJsonObject
 import net.primal.domain.nostr.utils.decodeLNUrlOrNull
 import net.primal.domain.nostr.utils.parseAsLNUrlOrNull
 import net.primal.domain.nostr.utils.parseHashtags
-import net.primal.domain.nostr.utils.parseNostrUris
+import net.primal.domain.nostr.utils.parseNostrUriCandidates
 
 fun List<NostrEvent>.mapAsProfileDataPO(
     cdnResources: List<CdnResource>,
@@ -100,7 +100,7 @@ fun NostrEvent.asProfileDataPO(
         lightningAddress = metadata?.lud16,
         lnUrlDecoded = metadata?.lud16?.parseAsLNUrlOrNull() ?: metadata?.lud06?.decodeLNUrlOrNull(),
         about = metadata?.about,
-        aboutUris = metadata?.about?.let { it.detectUrls() + it.parseNostrUris() } ?: emptyList(),
+        aboutUris = metadata?.about?.let { it.detectUrls() + it.parseNostrUriCandidates() } ?: emptyList(),
         aboutHashtags = metadata?.about?.parseHashtags() ?: emptyList(),
         displayName = metadata?.displayName,
         avatarCdnImage = metadata?.picture?.let {